refactor: move busines logic of all the instructions to the separate directory
