+++ fixed-price-sale/program/tests/init_selling_resource.rs
@@ -108,8 +108,8 @@ mod init_selling_resource {
         .to_account_metas(None);
 
         let data = mpl_fixed_price_sale_instruction::InitSellingResource {
-            _master_edition_bump: master_edition_bump,
-            _vault_owner_bump: vault_owner_bump,
+            master_edition_bump: master_edition_bump,
+            vault_owner_bump: vault_owner_bump,
             max_supply: Some(1),
         }
         .data();
@@ -241,8 +241,8 @@ mod init_selling_resource {
         .to_account_metas(None);
 
         let data = mpl_fixed_price_sale_instruction::InitSellingResource {
-            _master_edition_bump: master_edition_bump,
-            _vault_owner_bump: vault_owner_bump,
+            master_edition_bump: master_edition_bump,
+            vault_owner_bump: vault_owner_bump,
             max_supply: Some(1337),
         }
         .data();
@@ -364,8 +364,8 @@ mod init_selling_resource {
         .to_account_metas(None);
 
         let data = mpl_fixed_price_sale_instruction::InitSellingResource {
-            _master_edition_bump: master_edition_bump,
-            _vault_owner_bump: vault_owner_bump,
+            master_edition_bump: master_edition_bump,
+            vault_owner_bump: vault_owner_bump,
             max_supply: None,
         }
         .data();
